🌐 feat: create context & useContext introduced
