+++ src/components/city/CityList.jsx
@@ -4,8 +4,11 @@ import styles from './CityList.module.css';
 import Spinner from '../spinner/Spinner';
 import CityItem from './CityItem';
 import Message from '../message/Message';
+import { useCities } from '../../contexts/CitiesContext';
 
-const CityList = ({ cities, isLoading }) => {
+const CityList = () => {
+  const { cities, isLoading } = useCities();
+
   if (isLoading) return <Spinner />;
 
   if (!cities.length)
